Changed sortable scope target element
Changed sortable scope target element to the parent of 'placeElement'.

This fixes a bug where the placeHolder element would be appended to the wrong parent if the "as-sortable-item" element isn't a direct child of the 'as-sortable' element. This fixes flickering when using as-sortable on a table element while using a tbody or similar.

This might fix #96 and #132.